Move syncer built-in schema generation to dedicated package
Moves the syncer built-in schema generation to a dedicated package such
that it can be consumed in multiple locations. Updates the syncer API
reconciler to use it as first consumer.
